change int to double

+++ src/robocops_duplos/src/duplo_publisher.cpp
@@ -228,7 +228,7 @@ int main(int argc, char **argv)
             bool found = false;
             for (robocops_msgs::msg::Duplo &existing_duplo : duplos_buffer)
             {
-                RCLCPP_INFO(node->get_logger(), "Distance between %d and %d: %d", existing_duplo.id, new_duplo.id, calculate_distance(existing_duplo.position.point, new_duplo.position.point));
+                RCLCPP_INFO(node->get_logger(), "Distance between %d and %d: %d", existing_duplo.id, new_duplo.id, (int) calculate_distance(existing_duplo.position.point, new_duplo.position.point));
 
                 {
                     found = true;
